+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
+## [v3.3.4] - 2024-07-27
+### New Features
+- Added capability to setFilterDefaultValue for a DateRangeFilter by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1796
+- Add localised pill values for DateFilter, DateTimeFilter, DateRangeFilter by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1797
+
+### Tweaks
+- Migrating Carbon usage into Trait, Adding Filter/Search Lifecycle Hooks by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1798
+
 ## [v3.3.3] - 2024-07-23
 ### New Features
 - Add additional DateRangeFilter options by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1793